src/main/java/com/etrusted/interview/demo/service/UserServiceImpl.java
package com.etrusted.interview.demo.service;

import com.etrusted.interview.demo.entity.User;
import com.etrusted.interview.demo.repository.UserRepository;
import org.springframework.stereotype.Service;
import java.util.Optional;

@Service
public class UserServiceImpl implements UserService {

    private UserRepository userRepository;

    public UserServiceImpl(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    @Override
    public Optional<User> findUserByEmail(String email) {
        return userRepository.findUserByEmail(email);
    }

}
